python/requirements.txt
grpcio
grpcio-tools
rembg==2.0.59
